requirements.txt
pytorch
torchvision
torchaudio
pytorch-cuda==12.1
gradio
diffusers